requirements.txt
torch
torchvision
timm
numpy
opencv-python
scikit-learn
scikit-image
tqdm
pillow
matplotlib
grad-cam
lime